display: detect monitor border crossings if we jump over the edge
The granularity of the pointer updates varies. We might miss a border
crossing if the pointer moves more than one pixel in any direction. This
is problematic if a screen has more than one monitor which are
configured to be next to each other but are displayed in another way.
The pointer would move to the neighbor monitor of the system when it
should actually move to a completely other machine (which is displayed
next to the monitor).