usb: only send mouse position when we got a position response for the last position we sent
A new wey firmware always sends a positioning command when we triggered
a border crossing. This happens even when there is nothing behind that
border (e.g. the screen is displayed in the left bottom corner of the
physical monitor and we triggered the left border of the screen).

This helps with multi monitor systems where the X server itself handles
the cursor movement from one screen to another, but since only one of
the many screens can be displayed (e.g. bottom left corner) we need to
prevent that. Only the wey controller knows of the actual layout of the
screens displayed, it will reposition the cursor on the screen we just
left. However we may briefly move to the other screen and trigger its
border which confuses everyone. We now prevent this and only trigger a
border when we got a position command for the last border we triggered.