display: use XFIXES pointer barriers to make sure we can't leave the screen
We now can only leave the screen/monitor if the wey device told us so.
This fixes issues when the pointer was briefly visible on a neighbor
screen before the wey device had time to reposition the pointer. This
may happen when the displayed layout on the video wall doesn't match the
configuration of the layout in the X server config.