main,display,lmss: add cmd argument do disable drag'n'drop over internal borders
By default we disable the events generated when a monitor is left with a
mouse button pressed. This allows dragging of elements such as windows
across to a monitor next to the one we're leaving.

If the monitors are not displayed the way they are configured in XRandR
or Xorg, or it is generally not desired to be able to drag things around
an extended desktop, we can now disable this behavior by providing the
'-d' command line argument. With this option set, barrier events will be
generated regardless of the mouse button state.